refactor(tui): extract message-projection/pet-payload mixins from tui_gateway/server.py (shard s3)
