+++ tui_gateway/server.py
@@ -6692,382 +6692,6 @@ def _build_persist_user_message(user_text: str, image_paths: list[str], run_mess
     return [{"type": "text", "text": persist_text}, *image_parts]
 
 
-def _content_display_text(content: Any) -> str:
-    if content is None:
-        return ""
-    if isinstance(content, str):
-        return content
-    if isinstance(content, (int, float)):
-        return str(content)
-    if isinstance(content, list):
-        parts = []
-        for part in content:
-            text = _content_display_text(part).strip()
-            if text:
-                parts.append(text)
-        return "\n".join(parts)
-    if isinstance(content, dict):
-        kind = content.get("type")
-        if kind in {"text", "input_text", "output_text"}:
-            return str(content.get("text") or content.get("content") or "")
-        if kind in {"image_url", "input_image", "image"}:
-            return "[image]"
-        if kind in {"input_audio", "audio"}:
-            return "[audio]"
-        if kind:
-            return f"[{kind}]"
-        if "text" in content:
-            return str(content.get("text") or "")
-        return "[structured content]"
-    return str(content)
-
-
-def _coerce_message_text(content: Any) -> str:
-    """Render ``message['content']`` as a plain string for transport.
-
-    Provider-side, ``content`` may be a string (most common), a list of
-    multimodal parts (e.g. ``[{"type": "text", "text": "..."},
-    {"type": "image_url", "image_url": {...}}]``), or a single structured
-    dict. Calling ``.strip()`` on a list raises ``'list' object has no
-    attribute 'strip'`` and breaks session resume entirely.
-
-    Image parts (``image_url``) are preserved by appending the underlying
-    URL (data: or http:) into the text. The desktop renderer pulls these
-    back out via ``extractEmbeddedImages`` so the user sees the image
-    instead of the URL — and it stops the resume payload from disagreeing
-    with the cached message (which would otherwise cause the inline image
-    to flash, then disappear when the resume payload overwrites the cache).
-
-    Other structured dict shapes (audio, unknown types) fall back to a
-    bracketed placeholder so resume doesn't drop the message entirely.
-    """
-    if content is None:
-        return ""
-    if isinstance(content, str):
-        return content
-    if isinstance(content, (int, float)):
-        return str(content)
-    if isinstance(content, list):
-        chunks: list[str] = []
-        for part in content:
-            if isinstance(part, str):
-                chunks.append(part)
-                continue
-            if not isinstance(part, dict):
-                continue
-            text = part.get("text")
-            if isinstance(text, str):
-                chunks.append(text)
-                continue
-            kind = part.get("type")
-            if kind in {"text", "input_text", "output_text"}:
-                t = part.get("text") or part.get("content") or ""
-                if t:
-                    chunks.append(str(t))
-                continue
-            if kind in {"image_url", "input_image", "image"}:
-                image_url = part.get("image_url")
-                url = ""
-                if isinstance(image_url, dict):
-                    candidate = image_url.get("url")
-                    if isinstance(candidate, str):
-                        url = candidate
-                elif isinstance(image_url, str):
-                    url = image_url
-                if url:
-                    chunks.append(f"\n{url}")
-                else:
-                    chunks.append("\n[image]")
-                continue
-            if kind in {"input_audio", "audio"}:
-                chunks.append("\n[audio]")
-                continue
-            if kind:
-                chunks.append(f"\n[{kind}]")
-        return "".join(chunks)
-    if isinstance(content, dict):
-        kind = content.get("type")
-        if kind in {"text", "input_text", "output_text"}:
-            return str(content.get("text") or content.get("content") or "")
-        if kind in {"image_url", "input_image", "image"}:
-            image_url = content.get("image_url")
-            url = ""
-            if isinstance(image_url, dict):
-                candidate = image_url.get("url")
-                if isinstance(candidate, str):
-                    url = candidate
-            elif isinstance(image_url, str):
-                url = image_url
-            return url or "[image]"
-        if kind in {"input_audio", "audio"}:
-            return "[audio]"
-        if kind:
-            return f"[{kind}]"
-        if "text" in content:
-            return str(content.get("text") or "")
-        return "[structured content]"
-    return str(content)
-
-
-_TEXT_ONLY_BUSY_PART_KINDS = frozenset({"text", "input_text", "output_text"})
-
-
-def _is_text_only_busy_payload(content: Any) -> bool:
-    """True when a busy submit carries only plain text, not attachments/media."""
-    if content is None:
-        return False
-    if isinstance(content, (str, int, float)):
-        return True
-    if isinstance(content, list):
-        if not content:
-            return False
-        for part in content:
-            if isinstance(part, str):
-                continue
-            if not isinstance(part, dict):
-                return False
-            kind = part.get("type")
-            if kind in _TEXT_ONLY_BUSY_PART_KINDS:
-                continue
-            if kind is None and isinstance(part.get("text"), str):
-                continue
-            return False
-        return True
-    if isinstance(content, dict):
-        kind = content.get("type")
-        if kind in _TEXT_ONLY_BUSY_PART_KINDS:
-            return True
-        return kind is None and isinstance(content.get("text"), str)
-    return False
-
-
-def _is_display_hidden_marker(role: str | None, text: str) -> bool:
-    """Gateway bookkeeping notices (model-switch, personality) are persisted as
-    role=user ``[System: …]`` rows so strict providers accept them mid-history.
-    They are model-facing runtime metadata, not user turns, and must never
-    render as a user bubble in ANY client transcript (desktop, TUI, CLI, web).
-
-    Filtering here — the single display projection every surface reads — hides
-    them everywhere while the raw marker stays in ``session["history"]`` for the
-    model. It also removes the stored marker from the payload the desktop
-    reconciles against, so it can no longer shift user-message ordinals and
-    duplicate the optimistic prompt (#67603)."""
-    return role == "user" and text.lstrip().startswith("[System:")
-
-
-def _skill_scaffold_projection(content_text: str) -> str:
-    """Return the invocation a slash-skill-expanded turn came from, else "".
-
-    A ``/skill`` invocation expands into a model-facing message that embeds the
-    whole skill body. That payload belongs to the agent — every UI renders the
-    invocation (``/work fix the leak``) instead, so no surface can leak the
-    body into a chat bubble.
-    """
-    return describe_skill_invocation(content_text, separator=" ") or ""
-
-
-def _expand_skill_invocation_for_replay(text: str, task_id: str) -> str:
-    """Re-expand a projected `/skill` invocation before re-running that turn.
-
-    The inverse of :func:`_skill_scaffold_projection`. Because a skill turn is
-    displayed as its invocation, a rewind/regenerate hands us back
-    ``/work fix the leak`` rather than the body the agent originally saw —
-    re-running that verbatim would drop the skill. Re-expanding here keeps the
-    body server-side (no client ever holds it) and makes the replayed turn
-    identical to the original.
-
-    Returns *text* unchanged when it isn't a resolvable skill invocation.
-    """
-    head, _, arg = (text or "").strip().partition(" ")
-    if not head.startswith("/"):
-        return text
-
-    try:
-        from agent.skill_commands import (
-            build_skill_invocation_message,
-            resolve_skill_command_key,
-        )
-
-        cmd_key = resolve_skill_command_key(head.lstrip("/"))
-        if cmd_key is None:
-            return text
-
-        return build_skill_invocation_message(cmd_key, arg.strip(), task_id=task_id) or text
-    except Exception:
-        # A skill that no longer resolves (renamed, disabled, external dir
-        # gone) must not break the rewind — replay the text as typed.
-        logger.debug("skill re-expansion failed for replay", exc_info=True)
-        return text
-
-
-# Opening of the crash-recovery note synthesized by _auto_continue_note.
-# Matched (not just built) so a row persisted before the display type was
-# stamped at turn start still reads as a timeline event, and to recognize the
-# messaging gateway's twin note.
-_AUTO_CONTINUE_NOTE_PREFIX = "[System note: Your previous turn was interrupted mid-run"
-
-
-def _legacy_display_kind(role: str, text: str) -> str | None:
-    """Infer the display type of a synthetic row persisted without one.
-
-    Turn-start typing (see ``persist_user_display_kind``) covers everything
-    written from here on. Sessions already on disk carry untyped rows — and a
-    turn killed mid-run never reached the post-turn stamp at all, which is
-    exactly the auto-continue case — so the raw recovery note would paint as a
-    user bubble forever. Sniffing the one fixed synthetic prefix is the
-    migration for those rows; it is not how new rows get typed.
-    """
-    if role == "user" and text.lstrip().startswith(_AUTO_CONTINUE_NOTE_PREFIX):
-        return "auto_continue"
-    return None
-
-
-def _history_to_messages(history: list[dict]) -> list[dict]:
-    messages = []
-    tool_call_args = {}
-
-    for m in history:
-        if not isinstance(m, dict):
-            continue
-        role = m.get("role")
-        if role not in {"user", "assistant", "tool", "system"}:
-            continue
-        # An explicit display_kind="hidden" row is model-facing scaffolding
-        # (compaction references, interrupted-turn checkpoints). The string
-        # sniff below only catches the "[System:" convention; honor the
-        # declared field too, or scaffolding reaches every surface that reads
-        # this projection.
-        if m.get("display_kind") == "hidden":
-            continue
-        content_text = _coerce_message_text(m.get("content"))
-        if _is_display_hidden_marker(role, content_text):
-            continue
-        if role == "assistant" and m.get("tool_calls"):
-            for tc in m["tool_calls"]:
-                fn = tc.get("function", {})
-                tc_id = tc.get("id", "")
-                if tc_id and fn.get("name"):
-                    try:
-                        args = json.loads(fn.get("arguments", "{}"))
-                    except (json.JSONDecodeError, TypeError):
-                        args = {}
-                    tool_call_args[tc_id] = (fn["name"], args)
-            if not content_text.strip():
-                continue
-        if role == "tool":
-            tc_id = m.get("tool_call_id", "")
-            tc_info = tool_call_args.get(tc_id) if tc_id else None
-            name = (tc_info[0] if tc_info else None) or m.get("tool_name") or "tool"
-            args = (tc_info[1] if tc_info else None) or {}
-            messages.append(
-                {"role": "tool", "name": name, "context": _tool_ctx(name, args)}
-            )
-            continue
-        # An assistant turn may carry only reasoning/thinking content with no
-        # visible text (extended-thinking turns, thinking-only recovery
-        # responses). Such a turn is persisted with its reasoning fields and is
-        # recallable from the transcript, but dropping it here as "empty" makes
-        # it vanish from the resumed/reloaded session view while the desktop's
-        # reasoning disclosure has nothing to render. Keep it when it carries
-        # reasoning so the "Thinking…" block still shows. (#44022)
-        reasoning_keys = (
-            "reasoning",
-            "reasoning_content",
-            "reasoning_details",
-            "codex_reasoning_items",
-        )
-        has_reasoning = role == "assistant" and any(
-            m.get(key) for key in reasoning_keys
-        )
-        if not content_text.strip() and not has_reasoning:
-            continue
-        msg = {"role": role, "text": content_text}
-        # Durable row identity, stamped by _rows_to_conversation. The renderer's
-        # own message ids are ephemeral (timestamp+index derived, and a
-        # different shape for live vs rehydrated vs optimistic rows), so
-        # anything that addresses a specific persisted message later — message
-        # reactions — needs this instead.
-        if m.get("_row_id") is not None:
-            msg["row_id"] = m["_row_id"]
-        if role == "user":
-            invocation = _skill_scaffold_projection(content_text)
-            if invocation:
-                # Show the invocation, never the expanded skill body. The raw
-                # payload stays server-side: a rewind/regenerate re-sends the
-                # turn by ordinal, so no client needs it.
-                msg["text"] = invocation
-                msg["display_kind"] = "skill_invocation"
-        if role == "assistant":
-            for key in reasoning_keys:
-                if key in m and m.get(key) is not None:
-                    msg[key] = m.get(key)
-        # Forward display-only timeline metadata so the TUI can render
-        # model switches and delegation completions as events instead of
-        # opaque user messages, and hide compaction handoffs entirely.
-        display_kind = m.get("display_kind") or _legacy_display_kind(role, content_text)
-        if display_kind:
-            msg["display_kind"] = display_kind
-        if m.get("display_metadata"):
-            msg["display_metadata"] = m["display_metadata"]
-        messages.append(msg)
-
-    return messages
-
-
-def _coerce_seed_history(value: Any) -> list[dict]:
-    if not isinstance(value, list):
-        return []
-
-    history = []
-    for item in value:
-        if not isinstance(item, dict):
-            continue
-
-        role = item.get("role")
-        if role not in ("user", "assistant", "system"):
-            continue
-
-        content = item.get("content")
-        if content is None:
-            content = item.get("text")
-        if not isinstance(content, str) or not content.strip():
-            continue
-
-        history.append({"role": role, "content": content})
-
-    return history
-
-
-def _content_display_text(content: Any) -> str:
-    if content is None:
-        return ""
-    if isinstance(content, str):
-        return content
-    if isinstance(content, (int, float)):
-        return str(content)
-    if isinstance(content, list):
-        parts = []
-        for part in content:
-            text = _content_display_text(part).strip()
-            if text:
-                parts.append(text)
-        return "\n".join(parts)
-    if isinstance(content, dict):
-        kind = content.get("type")
-        if kind in {"text", "input_text", "output_text"}:
-            return str(content.get("text") or content.get("content") or "")
-        if kind in {"image_url", "input_image", "image"}:
-            return "[image]"
-        if kind in {"input_audio", "audio"}:
-            return "[audio]"
-        if kind:
-            return f"[{kind}]"
-        if "text" in content:
-            return str(content.get("text") or "")
-        return "[structured content]"
-    return str(content)
-
-
 def _inflight_text(value: Any) -> str:
     return _content_display_text(value).strip()
 
@@ -7981,302 +7605,6 @@ def _main_runtime_from_agent(agent) -> dict | None:
         elif field == "api_key" and callable(value):
             runtime[field] = value
     return runtime or None
-
-
-def _pet_frame_counts(spritesheet) -> dict:
-    """Real (padding-trimmed) frame count per state, for the desktop canvas.
-
-    Fail-open: a decode hiccup returns ``{}`` and the canvas falls back to its
-    static ``framesPerState`` rather than breaking the (cosmetic) pet.
-    """
-    try:
-        from agent.pet import render
-
-        return render.state_frame_counts(str(spritesheet))
-    except Exception:  # noqa: BLE001 - cosmetic, never break the surface
-        return {}
-
-
-_pet_payload_cache_lock = threading.Lock()
-_pet_payload_cache: dict[tuple, dict] = {}
-
-
-def _pet_sheet_revision(spritesheet) -> str:
-    """Stable revision id for one spritesheet file."""
-    try:
-        stat = spritesheet.stat()
-        return f"{stat.st_mtime_ns}:{stat.st_size}"
-    except Exception:  # noqa: BLE001 - cosmetic, never break the surface
-        return "0:0"
-
-
-def _pet_payload_cache_key(pet, *, scale: float) -> tuple | None:
-    """Cache key for the expensive sprite payload build."""
-    try:
-        stat = pet.spritesheet.stat()
-    except Exception:  # noqa: BLE001
-        return None
-    return (
-        str(pet.spritesheet),
-        stat.st_mtime_ns,
-        stat.st_size,
-        pet.slug,
-        pet.display_name,
-        round(scale, 4),
-    )
-
-
-def _clone_pet_payload(payload: dict) -> dict:
-    """Shallow-clone cached payloads so callers can't mutate shared state."""
-    out = dict(payload)
-    if isinstance(payload.get("framesByState"), dict):
-        out["framesByState"] = dict(payload["framesByState"])
-    if isinstance(payload.get("framesByRow"), dict):
-        out["framesByRow"] = dict(payload["framesByRow"])
-    if isinstance(payload.get("stateRows"), list):
-        out["stateRows"] = list(payload["stateRows"])
-    return out
-
-
-def _pet_row_frame_counts(spritesheet) -> dict:
-    """Real frame count per concrete spritesheet row name."""
-    try:
-        from PIL import Image
-
-        from agent.pet import constants, render
-
-        with Image.open(spritesheet) as opened:
-            image = opened.convert("RGBA")
-        cols = max(1, image.width // constants.FRAME_W)
-        row_count = max(1, image.height // constants.FRAME_H)
-        rows = constants.state_rows_for_grid(row_count)
-        out: dict[str, int] = {}
-        for row_idx, name in enumerate(rows[:row_count]):
-            top = row_idx * constants.FRAME_H
-            count = 0
-            for col in range(cols):
-                left = col * constants.FRAME_W
-                frame = image.crop((left, top, left + constants.FRAME_W, top + constants.FRAME_H))
-                if render._frame_is_blank(frame):
-                    break
-                count += 1
-            out[name] = count
-        return out
-    except Exception:  # noqa: BLE001 - cosmetic, never break the surface
-        return {}
-
-
-def _pet_config_scale() -> float:
-    """Configured ``display.pet.scale`` (or the engine default), never raises."""
-    from agent.pet import constants
-
-    try:
-        from hermes_cli.config import load_config
-
-        cfg = load_config()
-        display = cfg.get("display", {}) if isinstance(cfg.get("display"), dict) else {}
-        pet_cfg = display.get("pet", {}) if isinstance(display.get("pet"), dict) else {}
-        return float(pet_cfg.get("scale", constants.DEFAULT_SCALE) or constants.DEFAULT_SCALE)
-    except Exception:  # noqa: BLE001
-        return constants.DEFAULT_SCALE
-
-
-def _pet_sprite_payload(pet, *, scale: float) -> dict:
-    """Build the renderer payload (spritesheet bytes + geometry) for *pet*.
-
-    Shared by ``pet.info`` (the active mascot) and ``pet.hatch`` (the unadopted
-    preview) so both feed the desktop canvas / TUI from one shape.
-    """
-    import base64
-
-    from agent.pet import constants
-
-    cache_key = _pet_payload_cache_key(pet, scale=scale)
-    if cache_key is not None:
-        with _pet_payload_cache_lock:
-            cached = _pet_payload_cache.get(cache_key)
-        if cached is not None:
-            return _clone_pet_payload(cached)
-
-    raw = pet.spritesheet.read_bytes()
-    suffix = pet.spritesheet.suffix.lower()
-    mime = "image/png" if suffix == ".png" else "image/webp"
-    payload = {
-        "slug": pet.slug,
-        "displayName": pet.display_name,
-        "mime": mime,
-        "spritesheetBase64": base64.standard_b64encode(raw).decode("ascii"),
-        "spritesheetRevision": _pet_sheet_revision(pet.spritesheet),
-        "frameW": constants.FRAME_W,
-        "frameH": constants.FRAME_H,
-        "framesPerState": constants.FRAMES_PER_STATE,
-        "framesByState": _pet_frame_counts(pet.spritesheet),
-        "framesByRow": _pet_row_frame_counts(pet.spritesheet),
-        "loopMs": constants.LOOP_MS,
-        "scale": scale,
-        "stateRows": _pet_state_rows(pet.spritesheet),
-    }
-    if cache_key is not None:
-        with _pet_payload_cache_lock:
-            _pet_payload_cache[cache_key] = payload
-            while len(_pet_payload_cache) > 8:
-                _pet_payload_cache.pop(next(iter(_pet_payload_cache)))
-    return _clone_pet_payload(payload)
-
-
-def _pet_active_selection():
-    """Resolve configured active pet + scale from config."""
-    from agent.pet import constants, store
-
-    try:
-        from hermes_cli.config import load_config
-
-        cfg = load_config()
-        display = cfg.get("display", {}) if isinstance(cfg.get("display"), dict) else {}
-        pet_cfg = display.get("pet", {}) if isinstance(display.get("pet"), dict) else {}
-    except Exception:
-        pet_cfg = {}
-
-    enabled = bool(pet_cfg.get("enabled"))
-    configured_slug = str(pet_cfg.get("slug", "") or "")
-    pet = store.resolve_active_pet(configured_slug) if enabled else None
-    scale = float(pet_cfg.get("scale", constants.DEFAULT_SCALE) or constants.DEFAULT_SCALE)
-    return enabled, pet, scale
-
-
-def _pet_state_rows(spritesheet) -> list[str]:
-    """Row taxonomy for the concrete active pet sheet.
-
-    Hermes has to support both the legacy 8-row petdex atlas and the current
-    Codex/petdex 9-row atlas. The desktop canvas gets this list and indexes it
-    with the same `PetState` names the Python renderer uses.
-    """
-    try:
-        from PIL import Image
-
-        from agent.pet import constants
-
-        with Image.open(spritesheet) as image:
-            row_count = max(1, image.height // constants.FRAME_H)
-        return list(constants.state_rows_for_grid(row_count))
-    except Exception:  # noqa: BLE001 - cosmetic, never break the surface
-        from agent.pet import constants
-
-        return list(constants.STATE_ROWS)
-
-
-def _pet_gen_root():
-    """Profile-scoped staging dir for in-progress generation drafts."""
-    from hermes_constants import get_hermes_home
-
-    root = get_hermes_home() / "cache" / "pet-gen"
-    root.mkdir(parents=True, exist_ok=True)
-    return root
-
-
-def _pet_gen_sweep(root, *, max_age_s: float = 3600.0) -> None:
-    """Drop stale draft staging dirs so cache never grows unbounded."""
-    import shutil
-    import time
-
-    try:
-        now = time.time()
-        for child in root.iterdir():
-            if child.is_dir() and now - child.stat().st_mtime > max_age_s:
-                shutil.rmtree(child, ignore_errors=True)
-    except Exception as exc:  # noqa: BLE001 - cleanup is best-effort
-        logger.debug("pet-gen sweep failed: %s", exc)
-
-
-def _pet_png_data_uri(path, *, max_px: int = 160) -> str:
-    """Downscaled PNG data URI for a draft image (small preview payload)."""
-    import base64
-    import io
-
-    from PIL import Image
-
-    with Image.open(path) as opened:
-        img = opened.convert("RGBA")
-    img.thumbnail((max_px, max_px), Image.LANCZOS)
-    buf = io.BytesIO()
-    img.save(buf, format="PNG")
-    return "data:image/png;base64," + base64.standard_b64encode(buf.getvalue()).decode("ascii")
-
-
-# Cooperative cancellation for the heavy pet generation paths. The client's Stop
-# aborts its RPC immediately, but the worker-pool generation keeps running unless
-# told to stop — pet.cancel flips a token's flag, which generate_base_drafts /
-# hatch_pet poll between provider calls to skip work they haven't started.
-_pet_cancel_lock = threading.Lock()
-_pet_cancelled: set[str] = set()
-_PET_REFERENCE_MIME_EXT = {
-    "png": "png",
-    "jpeg": "jpg",
-    "jpg": "jpg",
-    "webp": "webp",
-    "gif": "gif",
-}
-try:
-    _PET_REFERENCE_MAX_BYTES = max(
-        1,
-        int(os.environ.get("HERMES_PET_REFERENCE_MAX_BYTES") or str(16 * 1024 * 1024)),
-    )
-except (TypeError, ValueError):
-    _PET_REFERENCE_MAX_BYTES = 16 * 1024 * 1024
-
-
-def _pet_reference_images_from_data_url(ref_raw: str, stage) -> list:
-    """Decode + validate a reference-image data URL into the stage dir."""
-    import base64
-    import binascii
-    import re as _re
-
-    match = _re.match(r"^data:image/([a-zA-Z0-9.+-]+);base64,(.*)$", ref_raw, _re.DOTALL)
-    if not match:
-        raise ValueError("invalid reference image format")
-
-    mime = match.group(1).lower()
-    ext = _PET_REFERENCE_MIME_EXT.get(mime)
-    if ext is None:
-        raise ValueError("unsupported reference image type")
-
-    payload = "".join(match.group(2).split())
-    approx = (len(payload) * 3) // 4
-    if approx > _PET_REFERENCE_MAX_BYTES:
-        raise ValueError("reference image too large")
-
-    try:
-        raw = base64.b64decode(payload, validate=True)
-    except (binascii.Error, ValueError) as exc:
-        raise ValueError("invalid reference image data") from exc
-
-    if len(raw) > _PET_REFERENCE_MAX_BYTES:
-        raise ValueError("reference image too large")
-
-    ref_path = stage / f"reference.{ext}"
-    ref_path.write_bytes(raw)
-    return [ref_path]
-
-
-def _pet_cancel_arm(token: str) -> None:
-    """Clear a stale cancel flag at the start of a generate/hatch run."""
-    with _pet_cancel_lock:
-        _pet_cancelled.discard(token)
-
-
-def _pet_cancel_request(token: str) -> None:
-    with _pet_cancel_lock:
-        _pet_cancelled.add(token)
-
-
-def _pet_is_cancelled(token: str) -> bool:
-    with _pet_cancel_lock:
-        return token in _pet_cancelled
-
-
-def _pet_cancel_release(token: str) -> None:
-    with _pet_cancel_lock:
-        _pet_cancelled.discard(token)
 
 
 # ===========================================================================
@@ -13903,3 +13231,48 @@ for _m in (
 ):
     _m.register(sys.modules[__name__])
 del _m
+# ── Split helper modules (wave-1 godfile extraction; see shard-plan.json) ──
+# Imported at the end of this module, mirroring the methods_* split above:
+# message_projection binds _tool_ctx (defined at 5140) from this module, so
+# it must be imported only after that global exists. Names are re-exported
+# here so in-file callers (the s2 pet watcher, _auto_continue_note, busy
+# submit policy, live-view reconciliation, s4 billing rows) and external
+# consumers (`from tui_gateway.server import _history_to_messages`, and the
+# HandlerRegistry rebinding for methods_*) keep resolving them.
+from .message_projection import (  # noqa: E402
+    _AUTO_CONTINUE_NOTE_PREFIX,
+    _coerce_message_text,
+    _coerce_seed_history,
+    _content_display_text,
+    _expand_skill_invocation_for_replay,
+    _history_to_messages,
+    _is_display_hidden_marker,
+    _is_text_only_busy_payload,
+    _legacy_display_kind,
+    _skill_scaffold_projection,
+)
+from .pet_payload_mixin import (  # noqa: E402
+    _PET_REFERENCE_MAX_BYTES,
+    _PET_REFERENCE_MIME_EXT,
+    _clone_pet_payload,
+    _pet_active_selection,
+    _pet_cancel_arm,
+    _pet_cancel_lock,
+    _pet_cancel_release,
+    _pet_cancel_request,
+    _pet_cancelled,
+    _pet_config_scale,
+    _pet_frame_counts,
+    _pet_gen_root,
+    _pet_gen_sweep,
+    _pet_is_cancelled,
+    _pet_payload_cache,
+    _pet_payload_cache_key,
+    _pet_payload_cache_lock,
+    _pet_png_data_uri,
+    _pet_reference_images_from_data_url,
+    _pet_row_frame_counts,
+    _pet_sheet_revision,
+    _pet_sprite_payload,
+    _pet_state_rows,
+)
